[PoS] restructure integration modules
