+++ apps/src/lib/node/ledger/protocol/mod.rs
@@ -8,7 +8,7 @@ use anoma::ledger::gas::{self, BlockGasMeter, VpGasMeter, VpsGas};
 use anoma::ledger::ibc::{self, Ibc};
 use anoma::ledger::native_vp::{self, NativeVp};
 use anoma::ledger::parameters::{self, ParametersVp};
-use anoma::ledger::pos::{self, PoS};
+use anoma::ledger::pos::{self, PosVP};
 use anoma::ledger::storage::write_log::WriteLog;
 use anoma::proto::{self, Tx};
 use anoma::types::address::{Address, InternalAddress};
@@ -36,7 +36,7 @@ pub enum Error {
     #[error("IBC native VP: {0}")]
     IbcNativeVpError(ibc::Error),
     #[error("PoS native VP: {0}")]
-    PosNativeVpError(pos::Error),
+    PosNativeVpError(pos::vp::Error),
     #[error("Parameters native VP: {0}")]
     ParametersNativeVpError(parameters::Error),
     #[error("Access to an internal address {0} is forbidden")]
@@ -219,7 +219,7 @@ fn execute_vps(
 
                     let accepted: Result<bool> = match internal_addr {
                         InternalAddress::PoS => {
-                            let pos = PoS { ctx };
+                            let pos = PosVP { ctx };
                             let result = pos
                                 .validate_tx(tx_data, keys, &verifiers_addr)
                                 .map_err(Error::PosNativeVpError);
